Moved alerts fetch to Dashboard.
