+++ assets/js/components/Dashboard/Dashboard.tsx
@@ -1,11 +1,31 @@
 import React, { ComponentType, useEffect, useState } from "react";
 import { Outlet } from "react-router";
 import "../../../css/screenplay.scss";
+import { Alert } from "../../models/alert";
 import { Place } from "../../models/place";
+import { ScreensByAlert } from "../../models/screensByAlert";
 import Sidebar from "./Sidebar";
+
+interface AlertsResponse {
+  alerts: Alert[];
+  screens_by_alert: ScreensByAlert;
+}
 
 const Dashboard: ComponentType = () => {
   const [places, setPlaces] = useState<Place[]>([]);
+  const [alerts, setAlerts] = useState<Alert[]>([]);
+  const [screensByAlertMap, setScreensByAlertMap] = useState<ScreensByAlert>(
+    {}
+  );
+
+  useEffect(() => {
+    fetch("/api/alerts")
+      .then((response) => response.json())
+      .then(({ alerts, screens_by_alert }: AlertsResponse) => {
+        setAlerts(alerts);
+        setScreensByAlertMap(screens_by_alert);
+      });
+  }, []);
 
   useEffect(() => {
     fetch("/api/dashboard")
@@ -19,7 +39,7 @@ const Dashboard: ComponentType = () => {
     <div className="screenplay-container">
       <Sidebar />
       <div className="page-content">
-        <Outlet context={{ places }} />
+        <Outlet context={{ places, alerts, screensByAlertMap }} />
       </div>
     </div>
   );
